+++ OnAFaimJPABack/src/main/java/dao/DAOProduit.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import model.Produit;
 
-public interface DAOProduit extends DAOGeneric<Produit,Integer> {
+public interface DAOProduit extends DAOGeneric<Produit,Long> {
 	
 	public List<Produit> findBySize(String size); 
 	public List<Produit> findByType(String type);
